+++ src/main/java/slipp/controller/UserCreateController.java
@@ -3,7 +3,7 @@ package slipp.controller;
 import slipp.exception.UserDuplicationException;
 import slipp.model.User;
 import slipp.service.UserCreateService;
-import webserver.http.domain.controller.Controller;
+import webserver.http.controller.Controller;
 import webserver.http.domain.request.Method;
 import webserver.http.domain.request.Request;
 import webserver.http.domain.response.Response;
