+++ src/main/java/slipp/controller/UserListController.java
@@ -2,8 +2,8 @@ package slipp.controller;
 
 import slipp.db.DataBase;
 import slipp.model.User;
-import webserver.http.domain.controller.Controller;
-import webserver.http.domain.controller.template.TemplateCompiler;
+import webserver.http.controller.Controller;
+import webserver.http.controller.template.TemplateCompiler;
 import webserver.http.domain.request.Method;
 import webserver.http.domain.request.Request;
 import webserver.http.domain.response.Response;
