+++ src/main/java/webserver/http/controller/StaticResourceController.java
@@ -1,4 +1,4 @@
-package webserver.http.domain.controller;
+package webserver.http.controller;
 
 import webserver.http.domain.ContentType;
 import webserver.http.domain.request.Method;
