+++ src/main/java/webserver/http/domain/Headers.java
@@ -1,5 +1,8 @@
 package webserver.http.domain;
 
+import webserver.http.domain.cookie.Cookie;
+import webserver.http.domain.cookie.Cookies;
+
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
