+++ src/main/java/webserver/http/domain/cookie/Cookie.java
@@ -1,4 +1,4 @@
-package webserver.http.domain;
+package webserver.http.domain.cookie;
 
 import java.util.Objects;
