+++ src/test/java/webserver/http/cookie/CookieTest.java
@@ -1,9 +1,10 @@
-package webserver.http.domain;
+package webserver.http.cookie;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import webserver.http.domain.cookie.Cookie;
 
 import static org.assertj.core.api.Assertions.assertThat;
